requirements.txt

aiofiles==24.1.0
aiohappyeyeballs==2.4.2
aiohttp==3.10.7
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
astor==0.8.1
asyncio==3.4.3
attrs==24.2.0
beautifulsoup4==4.12.3
black==23.7.0
bs4==0.0.2
certifi==2024.8.30
click==8.1.7
colorama==0.4.6
distro==1.9.0
esprima==4.0.1
flake8==6.1.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.6
httpx==0.24.1
idna==3.10
Jinja2==3.1.4
jiter==0.5.0
json5==0.9.25
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
mccabe==0.7.0
multidict==6.1.0
mypy-extensions==1.0.0
openai==1.12.0
azure-core==1.29.5
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
pycodestyle==2.11.1
pydantic==2.9.2
pydantic_core==2.23.4
pyflakes==3.1.0
python-dotenv==1.0.1
referencing==0.35.1
rpds-py==0.20.0
sniffio==1.3.1
soupsieve==2.6
tinycss2==1.3.0
tqdm==4.66.5
typing==3.7.4.3
typing_extensions==4.12.2
webencodings==0.5.1
yarl==1.13.1
